# docs/openapi.yml
openapi: 3.0.0
...
info:
  description: Codewind API
  version: 0.1.0
  title: Codewind API
  contact:
    name: API support
    email: example@ibm.com
  license:
    name: Codewind Developer License
    url: 'https://github.com/eclipse/codewind/blob/master/LICENSE'

# tags:
  # - name: users
  #   description: Operations available to regular users
  # - name: admin
  #   description: Operations available only to authenticated admins
paths:
...
  /api/v1/projects/{id}:
    get:
      summary: Get the project object for a single project
      description: Get the project object for the object specified by {id}
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      responses:
        200:
          description: The project was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        404:
          $ref: '#/components/responses/ProjectNotFound'
        500:
          description: Any other error
...
  /api/v1/projects/{id}/network:
    post:
      summary: Connect a project to another project
      description: >
        Pass a projects url to a project so its able to talk to it
      parameters:
        - name: id
          in: path
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      requestBody:
        description: JSON object containing details of the project to connect this project to
        content:
          application/json:
            schema:
              type: object
              required:
                - projectID
                - projectName
                - projectURL
                - connectionID
                - connectionURL
              properties:
                projectID:
                  schema:
                    $ref: '#/components/schemas/ProjectID'
                projectName:
                  schema:
                    $ref: '#/components/schemas/ProjectName'
                projectURL:
                  type: string
                connectionID:
                  type: string
                connectionURL:
                  type: string
      responses:
        200:
          description: Connection added
          content:
            text/html:
              schema:
                type: string
        404:
          description: Project not found
          content:
            text/html:
              schema:
                type: string
        500:
          description: Internal Error
    get:
      summary: Get all the connected projects for a project
      description: >
        An object containing connected projects
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      responses:
        200:
          description: Object containing connected projects
          content:
            application/json:
              schema:
                type: object
        404:
          description: if the project with id was not found
    delete:
      summary: Delete a projects network object
      description: >
        Used to quickly reset the network object in prototyping
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      responses:
        200:
          description: Object containing connected projects
          content:
            application/json:
              schema:
                type: object
        404:
          description: if the project with id was not found
components:
  schemas:
    Project:
      type: object
      required:
        - projectID
        - name
        - language
        - projectType
        - host
        - ports
      properties:
        projectID:
          $ref: '#/components/schemas/ProjectID'
        name:
          $ref: '#/components/schemas/ProjectName'
        language:
          $ref: '#/components/schemas/ProjectLanguage'
        projectType:
          $ref: '#/components/schemas/ProjectType'
        host:
          type: string
          example: localhost
        ports:
          type: object
          properties:
            exposedPort:
              type: string
              example: "32801"
            internalPort:
              type: string
              example: "3000"
            internalDebugPort:
              type: string
              example: "9229"
    ProjectID:
      type: string
      format: uuid
      example: d290f1ee-6c54-4b01-90e6-d701748f0851
      pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
    ProjectName:
      description: Name of project, cannot contain any of ", &, /, \\, ?, %, *, :, |, <, > or spaces
      type: string
      example: tobesproject
    ProjectLanguage:
      type: string
    ProjectType:
      description: the build type for the project. Built-in types are 'liberty', 'spring', 'swift', 'nodejs', and 'docker'.
      type: string
...
  responses:
...
    ProjectNotFound:
      description: Project not found
      content:
        text/html:
          schema:
            type: string